requirements.txt
Flask==0.10.1
Flask-Login==0.2.7
Flask-Mail==0.9.0
Flask-SQLAlchemy==1.0
Flask-WTF==0.9.3
Jinja2==2.7.1
MarkupSafe==0.18
PyYAML==3.10
SQLAlchemy==0.8.3
WTForms==1.0.5
Werkzeug==0.9.4
blinker==1.3
enum34==0.9.19
itsdangerous==0.23
wsgiref==0.1.2
